+++ AssettoServer/Server/Configuration/WeatherConfiguration.cs
@@ -10,12 +10,12 @@ namespace AssettoServer.Server.Configuration
     public class WeatherConfiguration
     {
         public string Graphics { get; internal set; }
-        public int BaseTemperatureAmbient { get; internal set; }
-        public int BaseTemperatureRoad { get; internal set; }
-        public int VariationAmbient { get; internal set; }
-        public int VariationRoad { get; internal set; }
-        public int WindBaseSpeedMin { get; internal set; }
-        public int WindBaseSpeedMax { get; internal set; }
+        public float BaseTemperatureAmbient { get; internal set; }
+        public float BaseTemperatureRoad { get; internal set; }
+        public float VariationAmbient { get; internal set; }
+        public float VariationRoad { get; internal set; }
+        public float WindBaseSpeedMin { get; internal set; }
+        public float WindBaseSpeedMax { get; internal set; }
         public int WindBaseDirection { get; internal set; }
         public int WindVariationDirection { get; internal set; }
         public WeatherFxParams WeatherFxParams { get; internal set; }
